requirements.txt
Flask==2.3.3
Flask-Login==0.6.3     
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.0.5
WTForms==3.0.1
Werkzeug==2.3.7        
email-validator==2.1.0 
python-dotenv==1.0.1
bcrypt==4.1.2
itsdangerous==2.1.2
Jinja2==3.1.4          
prettytable==3.10.0
requests==2.32.3       
